refactor(kube): move KubeService decorators to an .internal package (#546)
Establishes the jhelm internal-package convention ahead of the 1.0 API
freeze: with no JPMS, every public type is API unless excluded, so
genuinely-internal implementation classes move into `.internal`
sub-packages that document them as non-API.

First cut — the two KubeService decorators, which are public only because
the auto-config (in a sibling package) constructs them and reference nothing
in their old package:
- RetryableKubeService, ObservableKubeService -> ...kube.service.internal
- their tests move alongside (same-package access preserved)
- JhelmKubeAutoConfiguration + its test import from the new package
- package-info.java marks the package as unsupported/internal

jhelm-kube + jhelm-cli build green. Follow-up slices migrate the remaining
kube impls (HelmKubeService, AsyncHelmKubeService, KubernetesClientProvider,
ResourcePluralizer) and RepoHttpClientFactory / static factories in core,
then publish the per-module "what's stable" statement.
