+++ jhelm-kube/src/test/java/org/alexmond/jhelm/kube/JhelmKubeAutoConfigurationTest.java
@@ -7,8 +7,8 @@ import org.alexmond.jhelm.core.metrics.JhelmMetrics;
 import org.alexmond.jhelm.core.service.KubeService;
 import org.alexmond.jhelm.kube.service.HelmKubeService;
 import org.alexmond.jhelm.kube.service.KubeClient;
-import org.alexmond.jhelm.kube.service.ObservableKubeService;
-import org.alexmond.jhelm.kube.service.RetryableKubeService;
+import org.alexmond.jhelm.kube.service.internal.ObservableKubeService;
+import org.alexmond.jhelm.kube.service.internal.RetryableKubeService;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
