+++ jhelm-kube/src/test/java/org/alexmond/jhelm/kube/service/internal/ObservableKubeServiceTest.java
@@ -1,4 +1,4 @@
-package org.alexmond.jhelm.kube.service;
+package org.alexmond.jhelm.kube.service.internal;
 
 import java.util.List;
 import java.util.Optional;
